requirements.txt
asgiref==3.5.0
attrs==21.4.0
black==22.1.0
click==8.0.4
Django==4.0.3
django-js-asset==2.0.0
django-mptt==0.13.4
django-ninja==0.17.0
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.7
pytest==7.1.0
pytest-django==4.5.2
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
